+++ CHANGELOG.md
@@ -8,9 +8,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/bbx0/borgreport/compare/v0.2.0...HEAD) - 2024-10-13
+## [Unreleased](https://github.com/bbx0/borgreport/compare/v0.2.0...HEAD) - 2024-10-23
 
-TBD
+### Added
+
+- [OpenMetrics](https://github.com/prometheus/OpenMetrics/blob/main/specification/OpenMetrics.md) Text Exporter (Prometheus Metrics)
+- Option `--metrics-to` to save metrics in application/openmetrics-text format
 
 ## [0.2.0](https://github.com/bbx0/borgreport/compare/v0.1.0...v0.2.0) - 2024-10-13